third_party/WebKit/LayoutTests/fast/dom/shadow/querySelector-for-useragent-shadowroot-expected.txt


  
crbug.com/337616: test for querySelectorAll with /shadow/ and /shadow-deep/

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("form /shadow-deep/ *").length is 1
PASS document.querySelectorAll("form /shadow-deep/ *")[0].id is "input"
PASS document.querySelectorAll("form /shadow/ *").length is 0
PASS document.querySelectorAll("input /shadow-deep/ *").length is 0
PASS document.querySelectorAll("input /shadow/ *").length is 0
PASS document.querySelectorAll("details /shadow-deep/ *").length is 2
PASS document.querySelectorAll("details /shadow-deep/ *")[0].id is "summary"
PASS document.querySelectorAll("details /shadow-deep/ *")[1].id is "p"
PASS document.querySelectorAll("details /shadow/ *").length is 0
PASS document.querySelectorAll("summary /shadow-deep/ *").length is 0
PASS document.querySelectorAll("summary /shadow/ *").length is 0
PASS document.querySelectorAll("meter /shadow-deep/ *").length is 0
PASS document.querySelectorAll("meter /shadow/ *").length is 0
PASS document.querySelectorAll("progress /shadow-deep/ *").length is 0
PASS document.querySelectorAll("progress /shadow/ *").length is 0
PASS successfullyParsed is true

TEST COMPLETE
